requirements.txt
# Minimal requirements for vit_imagenette_stage1_muon_kfac.py
# Note: on Vast.ai, install a CUDA-matching PyTorch build if torch is not already present.
# Example:
#   pip install --index-url https://download.pytorch.org/whl/cu121 torch torchvision
# Then install the remaining requirements from this file.

numpy>=1.24
torch>=2.2
torchvision>=0.17
Pillow>=9.0
timm>=1.0
datasets>=2.19
